Memory: Implemented the Bytes utility.
This utility class allows for blitting structure types to and from an
opaque array of bytes. In the future, it will likely support further
manipulation, but for now it is purely an abstraction over C#'s
System.Runtime.InteropServices.Marshal system.

NOTE: Many of the features used in this class are unsafe or unstable!
You need to be certain or explicit about the layout of your structures!

Signed-off-by: Raven <raven@az-raven.com>